bug # 20868 -- enable/disable button "replace" based on search location
Replacing of found items in Script Editor is only available when Location option is 'Current Tab' or 'All Tabs' so the button "replace" of stack "revSEFind">Stack "main" should only be enabled then.  For all other Locations the "replace" button should be disabled.   Also when replace is not available and disabled the label of the button should be set to empty.  It can be set back to item 1 when the user changes back to location 'Current Tab' or 'All Tabs'.