[[ DataGrid2 ]] Allow swipe and hold
This patch allows a user to swipe and reveal the action button and
then hold the touch for a period of time. If they release the touch
while the button is fully exposed then it remains revealed. Previously
if they did not release the touch within the swipe detection time the
action button would be hidden.