Update revsecommoneditorbehavior.livecodescript
added testing for if tFirstLine is tLastLine, though the speed difference is not measurable, both version take 12 milliseconds on a 3600 line script. But I agree it is a cleaner way to do it.

Additionally I changed 
put (the number of chars of line 1 to tFirstLine - 1 of tNewText) + 1 into tFirstLineChar
from + 2 to +1. This also keeps the selection if the cursor is positioned before the first char.

As far as I understand charIndex works only on fields not variables. I did test and there is not speed improvement using charIndex on the field.