[[ Bug 21172 ]] Fix revMail on mobile
Adding the common library to mobile caused the revMail function to break.
`pass revMail` was added to the handler, but the engine did not support that method.

Replaced the `pass` with a call to `mobileComposeMail` to restore functionality

On iPhone this should be a good solution (looking at the source code).  The calls for revMail are much simpler for Android, but the end result should be the same.