[[ Bug 21245 ]] SE indent errors with inline block comments
There are 2 related errors in the lineStripComments / __StripComment handlers in the revsecommoneditorbehavior.livecodescript code.

The first error is that "/*" is assumed to run to the end of the current line.  Since it is possible to have multiple inline block comments on a single line, a recursive call to __StripBlockComment is used to remove these comments.

The second error is that the loop that goes through the current line does not exit when a true comment to the end of the line is encountered.  The loop uses quotes to ensure that it isn't capturing comment characters in a string, but fails to exit.  Added an additional return value (pass by reference variable) indicating comment status of the current chunk.