[[ Bug 21427 ]] Allow nesting of Data Grid Behaviors
If a developer needed to insert handlers to override any of the
default DG behavior, they were unable to do so using a behavior.
Updated the initialization code to be able to handle one or more nested
behaviors between the DG group and the DG library.

Script local `sResourceStack` created to retain the location of the library.
The value is initialized in the `_Initialize` handler.  All references to
`_ResourceStack()` were normalized and changed to the new variable (4 cases
used `stack _ResourceStack()` and the rest omitted `stack` which is included
in the returned value).  Cases where a local was declared for the value were
changed to use the script local.

The `_ResourceStack()` handler looks at the behavior of `me` first and if it
does not find an expected object (`button "Data Grid" of group "Behaviors"`)
then it will try to next level behavior until there are no more to try.