[[ Android Archs ]] Update standalone settings UI for android archs
This patch updates the standalone settings dialog for multiple android archs.
Specifically it replaces the `Build for Android` button with a group `Android Targets`
containing a checkbox button for each architecture. The individual buttons do not have
scripts but the group has the script:

```
on mouseUp
   if word 1 of the target is not "button" then
      exit mouseUp
   end if
   
   local tTarget
   put the short name of the target into tTarget
   updateAndroidArch word 2 of tTarget, the hilite of the target
end mouseUp
```

The `updateAndroidArch` handler is in the android card behavior script.