[Bug 20701] Make manual height in PI consistent (#2003)
* [Bug 20701] Make manual height in PI consistent

Makes setting the height in PI consistent. PI will either use manual height or max height of tab if smaller than manual height. custom properties now have a field "value" that is larger and increases with increaing the height of PI.

Furthermore it fixes the delay when vScrollbar is on or off for accomodating for vScrollbar regarding right side of controls.