[[ Bug 22137 ]] Fix conditional breakpoint evaluation
This patch ensures that conditional breakpoints that evaluate to false are reevaluated
if the same breakpoint is encountered again. For example, if a conditional breakpoint
is in a repeat loop it should be evaluated each iteration. A previous patch indended to
ensure that errors encountered while debugging a previous event were not re-entered
caused conditional breakpoints to not re-evaluate each iteration if no other breakpoint
was encountered.