[22205] Ensure the check for IDE stacks is accurate
This patch ensures that user stacks with names similar to the leaf name
of IDE library stacks (e.g. "drawing", "diff" etc) are no longer treated
as IDE stacks. The check if the stack name is among the lines of
revInternal__ListLoadedLibraries() was removed, as all the stacks included
in this list have either their _ideoverride set to true, or begin with "rev".
The only exception to this is tsNetLibURL, so for this reason its _ideoverride
property is set when this stack is loaded (i.e. in revInternal__LoadIfLibrary)