[[ Deployment Docs ]] Proposed changes to the heading structure to make sections easier to find
On LiveCode.com it's not been straight forward to find the iOS deployment guide or the Android deployment guide etc. All the information is in here but because we are using #### instead of ## for the <platform> settings sections they are not displaying in the right-hand menu on https://livecode.com/docs/9-5-0/deployment/deploying-your-application/

I am proposing changing some important section headers from #### to ## to make them easier to discover.