[[ Deployment Docs ]] Various tweaks (#2099)
* [[ Deployment Docs ]] Proposed changes to the heading structure to make sections easier to find

On LiveCode.com it's not been straight forward to find the iOS deployment guide or the Android deployment guide etc. All the information is in here but because we are using #### instead of ## for the <platform> settings sections they are not displaying in the right-hand menu on https://livecode.com/docs/9-5-0/deployment/deploying-your-application/

I am proposing changing some important section headers from #### to ## to make them easier to discover.

* [[ Deployment Docs ]] Make sections easier to find continued.

…
-Forgot to change the HTML5 heading to ## in the previous pull request. Made it in this one.-

Proposed changes to the heading structure to make sections easier to find

On LiveCode.com it's not been straight forward to find the iOS deployment guide or the Android deployment guide etc. All the information is in here but because we are using #### instead of ## for the settings sections they are not displaying in the right-hand menu on https://livecode.com/docs/9-5-0/deployment/deploying-your-application/

I am proposing changing some important section headers from #### to ## to make them easier to discover.

* [[ Deployment Docs ]] Proposing Changes to heading titles

The headers for the settings should be <Platform> Settings rather that <Platform> Deployment. I have put this proposed changes in this pull request

* [[ Deployment Docs ]] Proposing Changes to heading titles continued

Forgot to include HTML5 deployment name change alongside the changes made in LiveCodeSteven-patch-3. I have included it in this pull request.

* [[ Deployment Docs ]] Fixing a typo I spotted

Fixing a typo I spotted in this doc whilst making other changes.