Adding new item to revMenuBarStandardContextMenu
Every time I prototype a UI in LiveCode I always look for an arrange object layer menu item when I right-click. Mainly because this is present in other tools I use such as Adobe XD and Moqups. Looking for a coding challenge I decided to figure out how to do it myself. 

I have added the ability to Send to Back, Move Backward, Move Forward and Bring to front from an arrange menu item when you right-click on an object. I understand that you can code this or you can do it from the object menu item from the top menu bar, but I personally think this is a good place for those options to be, I always look for it here when I am prototyping something.