[[ Bug 22897 ]] Ensure script colorization is preserved on Big Sur
This patch ensures the script colorization is preserved on Big Sur
when using the default (Source Code Pro) script editor font. This bug
occured because of an experimental offending "SVG table with colored glyphs"
included in the font (https://github.com/adobe-fonts/source-code-pro/issues/250).
This patch replaces the existing Source Code Pro ttf files with a newer version
that does not suffer from this bug.