[[ Bug 23257 ]] Fix build error when "Allow Http Connections" is checked (#2161)
This patch fixes an android build error when "Allow Http Connections" is checked.
This happened because the button "Allow Http Connections" was part of the group "permissions",
thus the group scripts affected the button under specific circumstances. This patch places the
"Allow Http Connections" button outside the "permissions" group, so as it is not affected by code
that applies to the group.